Move RenderedTree to own file.
